@@ -83,21 +83,21 @@ class HomeScreen extends StatefulWidget {
 }
 
 class _HomeScreen extends State<HomeScreen>
-    with TickerProviderStateMixin, SingleTickerProviderStateMixin {
+    with SingleTickerProviderStateMixin {
   SharedPref sharedPref = SharedPref();
   Settings settings = new Settings();
 
-  BannerAd _bannerAd;
+  BannerAd? _bannerAd;
   bool _isBannerAdReady = false;
-  InterstitialAd _interstitialAd;
+  InterstitialAd? _interstitialAd;
   bool _isInterstitialAdReady = false;
 
   static GlobalKey<ScaffoldState> _scaffoldKey = new GlobalKey<ScaffoldState>();
   String url = "";
 
-  Uri _initialUri;
-  Uri _latestUri;
-  StreamSubscription _sub;
+  Uri? _initialUri;
+  Uri? _latestUri;
+  StreamSubscription? _sub;
   bool goToWeb = true;
 
   List<StreamSubscription<Position>> webViewGPSPositionStreams = [];
@@ -116,7 +116,7 @@ class _HomeScreen extends State<HomeScreen>
     buildNumber: 'Unknown',
   );
 
-  TabController tabController;
+  TabController? tabController;
   int _currentIndex = 0;
 
   @override
@@ -129,22 +129,22 @@ class _HomeScreen extends State<HomeScreen>
     tabController = new TabController(
       initialIndex: 0,
       length: widget.settings.tabNavigationEnable == "1"
-          ? widget.settings.tab.length
+          ? widget.settings.tab!.length
           : 1,
       vsync: this,
     );
-    tabController.addListener(_handleTabSelection);
+    tabController!.addListener(_handleTabSelection);
 
     if (widget.settings.adBanner == "1") {
       String adBannerId = Platform.isAndroid
-          ? widget.settings.admobKeyAdBanner
-          : widget.settings.admobKeyAdBannerIos;
+          ? widget.settings.admobKeyAdBanner!
+          : widget.settings.admobKeyAdBannerIos!;
       // TODO: Initialize _bannerAd
       _bannerAd = BannerAd(
         adUnitId: adBannerId,
         request: AdRequest(),
         size: AdSize.banner,
-        listener: AdListener(
+        listener: BannerAdListener(
           onAdLoaded: (_) {
             setState(() {
               _isBannerAdReady = true;
@@ -158,40 +158,42 @@ class _HomeScreen extends State<HomeScreen>
         ),
       );
 
-      _bannerAd.load();
+      _bannerAd!.load();
     }
 
     if (widget.settings.adInterstitial == "1") {
-      String adInterstitialId = Platform.isAndroid
+      String? adInterstitialId = Platform.isAndroid
           ? widget.settings.admobKeyAdInterstitial
           : widget.settings.admobKeyAdInterstitialIos;
-      _interstitialAd = InterstitialAd(
-        adUnitId: adInterstitialId,
-        request: AdRequest(),
-        listener: AdListener(
-          onAdLoaded: (_) {
-            _isInterstitialAdReady = true;
-          },
-          onAdFailedToLoad: (ad, err) {
-            print('Failed to load an interstitial ad: ${err.message}');
-            _isInterstitialAdReady = false;
-            ad.dispose();
-          },
-        ),
-      );
 
       Timer.periodic(
-          new Duration(seconds: int.parse(widget.settings.admobDealy)),
+          new Duration(seconds: int.parse(widget.settings.admobDealy!)),
           (timer) {
-        _interstitialAd?.load();
-        _interstitialAd?.show();
+        InterstitialAd.load(
+          adUnitId: adInterstitialId!,
+          request: AdRequest(),
+          adLoadCallback: InterstitialAdLoadCallback(
+            onAdLoaded: (ad) {
+              _isInterstitialAdReady = true;
+              _interstitialAd = ad;
+              ad.show();
+            },
+            onAdFailedToLoad: (err) {
+              print('Failed to load an interstitial ad: ${err.message}');
+              _isInterstitialAdReady = false;
+              // ad.dispose();
+            },
+          ),
+        );
+        // _interstitialAd?.load();
+        // _interstitialAd?.show();
       });
     }
   }
 
   void _handleIncomingLinks() {
     if (!kIsWeb) {
-      _sub = uriLinkStream.listen((Uri uri) async {
+      _sub = uriLinkStream.listen((Uri? uri) async {
         if (!mounted) return;
         print('got uri: $uri');
         setState(() {
@@ -216,7 +218,7 @@ class _HomeScreen extends State<HomeScreen>
             });
           }
         } else {
-          key0.currentState._webViewController?.loadUrl(
+          key0.currentState!._webViewController?.loadUrl(
             urlRequest: URLRequest(
               url: Uri.parse(link),
             ),
@@ -233,7 +235,7 @@ class _HomeScreen extends State<HomeScreen>
   }
 
   _handleTabSelection() {
-    setState(() => _currentIndex = tabController.index);
+    setState(() => _currentIndex = tabController!.index);
   }
 
   @override
@@ -301,10 +303,10 @@ class _HomeScreen extends State<HomeScreen>
           });
         }
       } else {
-        key0.currentState._webViewController?.loadUrl(
+        key0.currentState!._webViewController?.loadUrl(
           urlRequest: URLRequest(
             url: Uri.parse(
-              _oneSignalHelper.url,
+              _oneSignalHelper.url!,
             ),
           ),
         );
@@ -321,16 +323,16 @@ class _HomeScreen extends State<HomeScreen>
 
     return WillPopScope(
       onWillPop: () async {
-        return getCurrentKey().currentState.goBack();
+        return getCurrentKey().currentState!.goBack();
       },
       child: Container(
           decoration: BoxDecoration(color: HexColor("#f5f4f4")),
           padding: EdgeInsets.only(bottom: bottomPadding),
           child: Scaffold(
             key: _scaffoldKey,
-            appBar: _renderAppBar(context, settings),
-            drawer: (widget.settings.leftNavigationIcon.value == "icon_menu" ||
-                    widget.settings.rightNavigationIcon.value == "icon_menu")
+            appBar: _renderAppBar(context, settings) as PreferredSizeWidget?,
+            drawer: (widget.settings.leftNavigationIcon!.value == "icon_menu" ||
+                    widget.settings.rightNavigationIcon!.value == "icon_menu")
                 ? Drawer(
                     child: ListView(
                       padding: const EdgeInsets.all(0.0),
@@ -339,10 +341,10 @@ class _HomeScreen extends State<HomeScreen>
                             decoration: BoxDecoration(
                               gradient: LinearGradient(
                                 colors: <Color>[
-                                  themeProvider.isLightTheme
+                                  themeProvider.isLightTheme!
                                       ? HexColor(settings.firstColor)
                                       : themeProvider.darkTheme.primaryColor,
-                                  themeProvider.isLightTheme
+                                  themeProvider.isLightTheme!
                                       ? HexColor(settings.secondColor)
                                       : themeProvider.darkTheme.primaryColor,
                                 ],
@@ -356,19 +358,19 @@ class _HomeScreen extends State<HomeScreen>
                                     width: 70.0,
                                     height: 70.0,
                                     child: Image.network(
-                                      settings.logoHeaderUrl,
+                                      settings.logoHeaderUrl!,
                                     ),
                                   ),
                                   Padding(
                                     padding: EdgeInsets.only(top: 5),
-                                    child: Text(settings.title,
+                                    child: Text(settings.title!,
                                         overflow: TextOverflow.ellipsis,
                                         style: TextStyle(
                                             color: Colors.white, fontSize: 16)),
                                   ),
                                   Padding(
                                     padding: EdgeInsets.only(top: 5),
-                                    child: Text(settings.subTitle,
+                                    child: Text(settings.subTitle!,
                                         overflow: TextOverflow.ellipsis,
                                         style: TextStyle(
                                             color: Colors.white, fontSize: 14)),
@@ -378,7 +380,7 @@ class _HomeScreen extends State<HomeScreen>
                             )),
                         DrawerListTitle(
                             icon: Icons.home,
-                            text: I18n.current.home,
+                            text: I18n.current!.home,
                             onTap: () async {
                               if (widget.settings.tabNavigationEnable == "1") {
                                 if (goToWeb) {
@@ -398,22 +400,22 @@ class _HomeScreen extends State<HomeScreen>
                                   setState(() => goToWeb = true);
                                 }
                               } else {
-                                key0.currentState._webViewController?.loadUrl(
+                                key0.currentState!._webViewController?.loadUrl(
                                   urlRequest: URLRequest(
-                                    url: Uri.parse(settings.url),
+                                    url: Uri.parse(url),
                                   ),
                                 );
 
                                 Navigator.pop(context);
                               }
                             }),
-                        _renderMenuDrawer(settings.menus, context),
+                        _renderMenuDrawer(settings.menus!, context),
                         Padding(
                           padding: EdgeInsets.fromLTRB(0, 20, 0, 20),
                           child: Divider(height: 1, color: Colors.grey[400]),
                         ),
-                        _renderPageDrawer(settings.pages, context),
-                        settings.pages.length != 0
+                        _renderPageDrawer(settings.pages!, context),
+                        settings.pages!.length != 0
                             ? Padding(
                                 padding: EdgeInsets.fromLTRB(0, 20, 0, 20),
                                 child:
@@ -422,11 +424,11 @@ class _HomeScreen extends State<HomeScreen>
                             : Container(height: 0),
                         DrawerListTitle(
                             icon: Icons.brightness_medium,
-                            text: themeProvider.isLightTheme
-                                ? I18n.current.darkMode
-                                : I18n.current.lightMode,
+                            text: themeProvider.isLightTheme!
+                                ? I18n.current!.darkMode
+                                : I18n.current!.lightMode,
                             onTap: () {
-                              if (themeProvider.isLightTheme) {
+                              if (themeProvider.isLightTheme!) {
                                 themeProvider.setDarkMode();
                               } else {
                                 themeProvider.setLightMode();
@@ -434,7 +436,7 @@ class _HomeScreen extends State<HomeScreen>
                             }),
                         DrawerListTitle(
                           icon: Icons.translate,
-                          text: I18n.current.languages,
+                          text: I18n.current!.languages,
                           onTap: () {
                             Navigator.pop(context);
                             Navigator.push(
@@ -462,7 +464,7 @@ class _HomeScreen extends State<HomeScreen>
                         ),
                         DrawerListTitle(
                           icon: Icons.info,
-                          text: I18n.current.about,
+                          text: I18n.current!.about,
                           onTap: () {
                             Navigator.pop(context);
                             Navigator.push(
@@ -476,13 +478,14 @@ class _HomeScreen extends State<HomeScreen>
                         ),
                         DrawerListTitle(
                             icon: Icons.share,
-                            text: I18n.current.share,
+                            text: I18n.current!.share,
                             onTap: () {
-                              shareApp(context, settings.title, settings.share);
+                              shareApp(
+                                  context, settings.title, settings.share!);
                             }),
                         DrawerListTitle(
                             icon: Icons.star,
-                            text: I18n.current.rate,
+                            text: I18n.current!.rate,
                             onTap: () => LaunchReview.launch(
                                 androidAppId: settings.androidId,
                                 iOSAppId: settings.iosId)),
@@ -498,7 +501,7 @@ class _HomeScreen extends State<HomeScreen>
                                 ),
                                 DrawerListTitle(
                                     icon: Icons.system_update,
-                                    text: I18n.current.update,
+                                    text: I18n.current!.update,
                                     onTap: () => LaunchReview.launch(
                                         androidAppId: settings.androidId,
                                         iOSAppId: settings.iosId))
@@ -518,10 +521,10 @@ class _HomeScreen extends State<HomeScreen>
                             controller: tabController,
                             physics: NeverScrollableScrollPhysics(),
                             children: List.generate(
-                              widget.settings.tab.length,
+                              widget.settings.tab!.length,
                               (index) => WebViewScreen(
                                   key: listKey[index],
-                                  path: widget.settings.tab[index].url,
+                                  path: widget.settings.tab![index].url,
                                   pos: index,
                                   settings: widget.settings,
                                   stream: listStream[index].stream),
@@ -538,7 +541,7 @@ class _HomeScreen extends State<HomeScreen>
                   if (widget.settings.adBanner == "1" && _isBannerAdReady)
                     Container(
                       height: 50,
-                      child: AdWidget(ad: _bannerAd),
+                      child: AdWidget(ad: _bannerAd!),
                     )
                 ]),
               ],
@@ -563,14 +566,22 @@ class _HomeScreen extends State<HomeScreen>
                         child: _buildTabItem(context)),
                   )
                 : Container(height: 0),
-            floatingActionButton: widget.settings.floating.length != 0
+            // floatingActionButtonLocation: FloatingActionButtonLocation.endFloat,
+            floatingActionButton: widget.settings.floating!.length != 0
                 ? SpeedDial(
                     icon: Icons.add,
                     backgroundColor: HexColor(widget.settings.firstColor),
                     foregroundColor: Colors.white,
                     children:
-                        _renderFloating(widget.settings.floating, context))
-                : null,
+                        _renderFloating(widget.settings.floating!, context),
+                  )
+                : Padding(
+                    padding: const EdgeInsets.only(bottom: 40),
+                    child: FloatingActionButton(
+                      onPressed: () => _scaffoldKey.currentState?.openDrawer(),
+                      child: Icon(Icons.menu),
+                    ),
+                  ),
             /*bottomNavigationBar: Container(
                 height: settings.adBanner == "1"
                     ? Platform.isAndroid
@@ -594,7 +605,7 @@ class _HomeScreen extends State<HomeScreen>
         },
         indicator: UnderlineTabIndicator(
           borderSide: BorderSide(
-              color: themeProvider.isLightTheme
+              color: themeProvider.isLightTheme!
                   ? tabColor
                   : themeProvider.darkTheme.primaryColor,
               width: 2.5),
@@ -603,29 +614,29 @@ class _HomeScreen extends State<HomeScreen>
         controller: tabController,
         labelColor: tabColor,
         unselectedLabelColor: Colors.black26,
-        tabs: List.generate(widget.settings.tab.length, (index) {
+        tabs: List.generate(widget.settings.tab!.length, (index) {
           return new Tab(
             child: new Column(
                 mainAxisAlignment: MainAxisAlignment.center,
                 crossAxisAlignment: CrossAxisAlignment.center,
                 children: <Widget>[
-                  new Image.network(settings.tab[index].icon_url,
+                  new Image.network(settings.tab![index].icon_url!,
                       width: 25,
                       height: 25,
                       color: _currentIndex == index
-                          ? themeProvider.isLightTheme
+                          ? themeProvider.isLightTheme!
                               ? tabColor
                               : themeProvider.darkTheme.primaryColor
                           : unselectedColor),
                   new SizedBox(height: 5),
                   new Flexible(
-                      child: new Text(settings.tab[index].title,
+                      child: new Text(settings.tab![index].title!,
                           overflow: TextOverflow.ellipsis,
                           style: TextStyle(
                               fontSize:
-                                  widget.settings.tab.length == 5 ? 8 : 10,
+                                  widget.settings.tab!.length == 5 ? 8 : 10,
                               color: _currentIndex == index
-                                  ? themeProvider.isLightTheme
+                                  ? themeProvider.isLightTheme!
                                       ? tabColor
                                       : themeProvider.darkTheme.primaryColor
                                   : unselectedColor))),
@@ -662,7 +673,7 @@ class _HomeScreen extends State<HomeScreen>
                           )),
                       SizedBox(height: 40),
                       Text(
-                        I18n.current.whoops,
+                        I18n.current!.whoops,
                         style: TextStyle(
                             color: Colors.black45,
                             fontSize: 40.0,
@@ -670,14 +681,14 @@ class _HomeScreen extends State<HomeScreen>
                       ),
                       SizedBox(height: 20),
                       Text(
-                        I18n.current.noInternet,
+                        I18n.current!.noInternet,
                         style: TextStyle(color: Colors.black87, fontSize: 15.0),
                       ),
                       SizedBox(height: 5),
                       SizedBox(height: 60),
                       RaisedGradientButton(
                           child: Text(
-                            I18n.current.tryAgain,
+                            I18n.current!.tryAgain,
                             style: TextStyle(
                                 color: Colors.white,
                                 fontSize: 18.0,
@@ -712,7 +723,7 @@ class _HomeScreen extends State<HomeScreen>
   }
    */
 
-  int parseInt(dynamic value) {
+  int? parseInt(dynamic value) {
     if (value == null) return null;
     if (value is int) return value;
 
@@ -723,25 +734,25 @@ class _HomeScreen extends State<HomeScreen>
       PositionOptions positionOptions) async {
     PositionResponse positionResponse = PositionResponse();
 
-    int timeout = 30000;
-    if (positionOptions.timeout > 0) timeout = positionOptions.timeout;
+    int? timeout = 30000;
+    if (positionOptions.timeout! > 0) timeout = positionOptions.timeout;
 
     try {
-      Geolocator geolocator = Geolocator()..forceAndroidLocationManager = true;
-      GeolocationStatus geolocationStatus =
-          await geolocator.checkGeolocationPermissionStatus();
+      // Geolocator geolocator = Geolocator()..forceAndroidLocationManager = true;
+      LocationPermission geolocationStatus =
+          await GeolocatorPlatform.instance.requestPermission();
 
-      if (geolocationStatus == GeolocationStatus.denied ||
-          geolocationStatus == GeolocationStatus.granted ||
-          geolocationStatus == GeolocationStatus.restricted ||
-          geolocationStatus == GeolocationStatus.unknown) {
+      if (geolocationStatus == LocationPermission.always ||
+          geolocationStatus == LocationPermission.whileInUse) {
         positionResponse.position = await Future.any([
-          geolocator.getCurrentPosition(
-              desiredAccuracy: (positionOptions.enableHighAccuracy
-                  ? LocationAccuracy.best
-                  : LocationAccuracy.medium)),
-          Future.delayed(Duration(milliseconds: timeout), () {
-            if (positionOptions.timeout > 0) positionResponse.timedOut = true;
+          GeolocatorPlatform.instance.getCurrentPosition(
+            locationSettings: LocationSettings(
+                accuracy: (positionOptions.enableHighAccuracy
+                    ? LocationAccuracy.best
+                    : LocationAccuracy.medium)),
+          ),
+          Future.delayed(Duration(milliseconds: timeout!), () {
+            if (positionOptions.timeout! > 0) positionResponse.timedOut = true;
             return;
           })
         ]);
@@ -780,7 +791,7 @@ class _HomeScreen extends State<HomeScreen>
       children: menus
           .map(
             (Menu menu) => DrawerListTitle(
-                icon_url: menu.iconUrl,
+                iconUrl: menu.iconUrl,
                 text: menu.title,
                 onTap: () async {
                   if (widget.settings.tabNavigationEnable == "1") {
@@ -804,8 +815,8 @@ class _HomeScreen extends State<HomeScreen>
                       });
                     }
                   } else {
-                    key0.currentState._webViewController?.loadUrl(
-                        urlRequest: URLRequest(url: Uri.parse(menu.url)));
+                    key0.currentState!._webViewController?.loadUrl(
+                        urlRequest: URLRequest(url: Uri.parse(menu.url!)));
 
                     Navigator.pop(context);
                   }
@@ -821,7 +832,7 @@ class _HomeScreen extends State<HomeScreen>
           (Floating floating) => SpeedDialChild(
               child: Container(
                 padding: EdgeInsets.all(13.0),
-                child: Image.network(floating.iconUrl,
+                child: Image.network(floating.iconUrl!,
                     width: 15, height: 15, color: HexColor(floating.iconColor)),
               ),
               label: floating.title,
@@ -845,9 +856,9 @@ class _HomeScreen extends State<HomeScreen>
                     setState(() => goToWeb = true);
                   }
                 } else {
-                  key0.currentState._webViewController?.loadUrl(
+                  key0.currentState!._webViewController?.loadUrl(
                     urlRequest: URLRequest(
-                      url: Uri.parse(floating.url),
+                      url: Uri.parse(floating.url!),
                     ),
                   );
 
@@ -859,21 +870,21 @@ class _HomeScreen extends State<HomeScreen>
   }
 
   Widget _renderPageDrawer(List<Page> pages, context) {
+    print(pages.map((e) => e.icon));
     return new Column(
       children: pages
-          .map(
-            (Page page) => DrawerListTitle(
-                icon_url: page.iconUrl,
-                text: page.title,
-                onTap: () {
-                  Navigator.pop(context);
-                  Navigator.push(
-                      context,
-                      PageTransition(
-                          type: pageTransitionAnimation(context),
-                          child: PageScreen(page, widget.settings)));
-                }),
-          )
+          .map((Page page) => DrawerListTitle(
+              forceThemeColor: true,
+              iconUrl: page.iconUrl,
+              text: page.title,
+              onTap: () {
+                Navigator.pop(context);
+                Navigator.push(
+                    context,
+                    PageTransition(
+                        type: pageTransitionAnimation(context),
+                        child: PageScreen(page, widget.settings)));
+              }))
           .toList(),
     );
   }
@@ -888,7 +899,7 @@ class _HomeScreen extends State<HomeScreen>
                 children: <Widget>[
                   _renderMenuIcon(
                       context,
-                      widget.settings.leftNavigationIcon,
+                      widget.settings.leftNavigationIcon!,
                       widget.settings.rightNavigationIcon,
                       widget.settings.navigatinBarStyle,
                       widget.settings,
@@ -898,7 +909,7 @@ class _HomeScreen extends State<HomeScreen>
                   Row(
                       children: _renderMenuIconList(
                           context,
-                          widget.settings.rightNavigationIconList,
+                          widget.settings.rightNavigationIconList!,
                           widget.settings.leftNavigationIcon,
                           widget.settings.navigatinBarStyle,
                           widget.settings,
@@ -911,10 +922,10 @@ class _HomeScreen extends State<HomeScreen>
                   begin: Alignment.centerLeft,
                   end: Alignment.centerRight,
                   colors: <Color>[
-                    themeProvider.isLightTheme
+                    themeProvider.isLightTheme!
                         ? HexColor(settings.firstColor)
                         : themeProvider.darkTheme.primaryColor,
-                    themeProvider.isLightTheme
+                    themeProvider.isLightTheme!
                         ? HexColor(settings.secondColor)
                         : themeProvider.darkTheme.primaryColor,
                   ],
@@ -928,7 +939,7 @@ class _HomeScreen extends State<HomeScreen>
             ));
   }
 
-  Widget _renderTitle(String type, Settings settings) {
+  Widget _renderTitle(String? type, Settings settings) {
     var direction = MainAxisAlignment.start;
 
     switch (type) {
@@ -953,7 +964,7 @@ class _HomeScreen extends State<HomeScreen>
             child: Container(
               child: settings.typeHeader == "text"
                   ? Text(
-                      settings.title,
+                      settings.title!,
                       overflow: TextOverflow.ellipsis,
                       style: TextStyle(
                           color: Colors.white,
@@ -961,7 +972,7 @@ class _HomeScreen extends State<HomeScreen>
                           fontWeight: FontWeight.bold),
                     )
                   : settings.typeHeader == "image"
-                      ? Image.network(settings.logoHeaderUrl, height: 40)
+                      ? Image.network(settings.logoHeaderUrl!, height: 40)
                       : Container(),
             ),
           )
@@ -973,8 +984,8 @@ class _HomeScreen extends State<HomeScreen>
   Widget _renderMenuIcon(
       BuildContext context,
       NavigationIcon navigationIcon,
-      NavigationIcon navigationOtherIcon,
-      String navigatinBarStyle,
+      NavigationIcon? navigationOtherIcon,
+      String? navigatinBarStyle,
       Settings settings,
       String direction) {
     return navigationIcon.value != "icon_empty"
@@ -989,10 +1000,10 @@ class _HomeScreen extends State<HomeScreen>
                       icon: Transform(
                           alignment: Alignment.center,
                           transform: Matrix4.rotationY(math.pi *
-                              (I18n.current.textDirection == TextDirection.ltr
+                              (I18n.current!.textDirection == TextDirection.ltr
                                   ? 2
                                   : 1)),
-                          child: new Image.network(navigationIcon.iconUrl,
+                          child: new Image.network(navigationIcon.iconUrl!,
                               height: 25, width: 25, color: Colors.white)
                           /*Image.asset(
                               UIImages.imageDir +
@@ -1009,7 +1020,7 @@ class _HomeScreen extends State<HomeScreen>
                     ),
                     Container(
                       width: (navigatinBarStyle == "center" &&
-                              navigationOtherIcon.value == "icon_back_forward")
+                              navigationOtherIcon!.value == "icon_back_forward")
                           ? 50
                           : 0,
                     )
@@ -1022,7 +1033,8 @@ class _HomeScreen extends State<HomeScreen>
                         icon: Transform(
                             alignment: Alignment.center,
                             transform: Matrix4.rotationY(math.pi *
-                                (I18n.current.textDirection == TextDirection.ltr
+                                (I18n.current!.textDirection ==
+                                        TextDirection.ltr
                                     ? 2
                                     : 1)),
                             child: Image.asset(
@@ -1034,29 +1046,29 @@ class _HomeScreen extends State<HomeScreen>
                           switch (_currentIndex) {
                             case 0:
                               {
-                                key0.currentState._webViewController?.goBack();
+                                key0.currentState!._webViewController?.goBack();
                               }
                               break;
 
                             case 1:
                               {
-                                key1.currentState._webViewController?.goBack();
+                                key1.currentState!._webViewController?.goBack();
                               }
                               break;
 
                             case 2:
                               {
-                                key2.currentState._webViewController?.goBack();
+                                key2.currentState!._webViewController?.goBack();
                               }
                               break;
                             case 3:
                               {
-                                key3.currentState._webViewController?.goBack();
+                                key3.currentState!._webViewController?.goBack();
                               }
                               break;
                             case 4:
                               {
-                                key4.currentState._webViewController?.goBack();
+                                key4.currentState!._webViewController?.goBack();
                               }
                               break;
                             default:
@@ -1072,7 +1084,8 @@ class _HomeScreen extends State<HomeScreen>
                         icon: Transform(
                             alignment: Alignment.center,
                             transform: Matrix4.rotationY(math.pi *
-                                (I18n.current.textDirection == TextDirection.ltr
+                                (I18n.current!.textDirection ==
+                                        TextDirection.ltr
                                     ? 2
                                     : 1)),
                             child: Image.asset(
@@ -1082,7 +1095,7 @@ class _HomeScreen extends State<HomeScreen>
                                 color: Colors.white)),
                         onPressed: () {
                           getCurrentKey()
-                              .currentState
+                              .currentState!
                               ._webViewController
                               ?.goForward();
                         },
@@ -1098,8 +1111,8 @@ class _HomeScreen extends State<HomeScreen>
   List<Widget> _renderMenuIconList(
       BuildContext context,
       List<NavigationIcon> navigationIcon,
-      NavigationIcon navigationOtherIcon,
-      String navigatinBarStyle,
+      NavigationIcon? navigationOtherIcon,
+      String? navigatinBarStyle,
       Settings settings,
       String direction) {
     return navigationIcon
@@ -1118,9 +1131,12 @@ class _HomeScreen extends State<HomeScreen>
 
   Future<bool> _onBackPressed(context) async {
     try {
-      if (getCurrentKey().currentState._webViewController != null) {
-        if (await getCurrentKey().currentState._webViewController.canGoBack()) {
-          getCurrentKey().currentState._webViewController.goBack();
+      if (getCurrentKey().currentState!._webViewController != null) {
+        if (await getCurrentKey()
+            .currentState!
+            ._webViewController!
+            .canGoBack()) {
+          getCurrentKey().currentState!._webViewController!.goBack();
           return false;
         } else {
           _showDialog(context);
@@ -1129,23 +1145,25 @@ class _HomeScreen extends State<HomeScreen>
     } catch (e) {
       _showDialog(context);
     }
+
+    return true;
   }
 
   _showDialog(context) {
     return showDialog(
       context: context,
       builder: (context) => new AlertDialog(
-        title: new Text(I18n.current.closeApp),
-        content: new Text(I18n.current.sureCloseApp),
+        title: new Text(I18n.current!.closeApp),
+        content: new Text(I18n.current!.sureCloseApp),
         actions: <Widget>[
           new FlatButton(
             onPressed: () => Navigator.of(context).pop(false),
-            child: new Text(I18n.current.cancel),
+            child: new Text(I18n.current!.cancel),
           ),
           SizedBox(height: 16),
           new FlatButton(
             onPressed: () => exit(0),
-            child: new Text(I18n.current.ok),
+            child: new Text(I18n.current!.ok),
           ),
         ],
       ),
@@ -1172,13 +1190,13 @@ class _HomeScreen extends State<HomeScreen>
           });
         }
       } else {
-        key0.currentState._webViewController?.loadUrl(
-            urlRequest: URLRequest(url: Uri.parse(navigationIcon.url)));
+        key0.currentState!._webViewController?.loadUrl(
+            urlRequest: URLRequest(url: Uri.parse(navigationIcon.url!)));
       }
     } else {
       switch (navigationIcon.value) {
         case "icon_menu":
-          _HomeScreen._scaffoldKey.currentState.openDrawer();
+          _HomeScreen._scaffoldKey.currentState!.openDrawer();
           break;
         case "icon_home":
           if (widget.settings.tabNavigationEnable == "1") {
@@ -1196,21 +1214,21 @@ class _HomeScreen extends State<HomeScreen>
               });
             }
           } else {
-            key0.currentState._webViewController
-                ?.loadUrl(urlRequest: URLRequest(url: Uri.parse(settings.url)));
+            key0.currentState!._webViewController?.loadUrl(
+                urlRequest: URLRequest(url: Uri.parse(settings.url!)));
           }
           break;
         case "icon_reload":
-          getCurrentKey().currentState._webViewController?.reload();
+          getCurrentKey().currentState!._webViewController?.reload();
           break;
         case "icon_share":
-          shareApp(context, settings.title, settings.share);
+          shareApp(context, settings.title, settings.share!);
           break;
         case "icon_back":
-          getCurrentKey().currentState._webViewController?.goBack();
+          getCurrentKey().currentState!._webViewController?.goBack();
           break;
         case "icon_forward":
-          getCurrentKey().currentState._webViewController?.goForward();
+          getCurrentKey().currentState!._webViewController?.goForward();
           break;
         case "icon_exit":
           _showDialog(context);
@@ -1237,14 +1255,14 @@ class _HomeScreen extends State<HomeScreen>
       if (!mounted) return;
       print("qrCode" + qrCode);
       setState(() {
-        key0.currentState.isLoading = true;
+        key0.currentState!.isLoading = true;
       });
 
       final response = await http.head(Uri.parse(qrCode));
 
       if (response.statusCode == 200) {
         setState(() {
-          key0.currentState.isLoading = false;
+          key0.currentState!.isLoading = false;
         });
         if (widget.settings.tabNavigationEnable == "1") {
           if (goToWeb) {
@@ -1262,24 +1280,24 @@ class _HomeScreen extends State<HomeScreen>
             });
           }
         } else {
-          key0.currentState._webViewController
+          key0.currentState!._webViewController
               ?.loadUrl(urlRequest: URLRequest(url: Uri.parse(qrCode)));
         }
       } else {
         setState(() {
-          key0.currentState.isLoading = false;
+          key0.currentState!.isLoading = false;
         });
       }
     } finally {
       setState(() {
-        key0.currentState.isLoading = false;
+        key0.currentState!.isLoading = false;
       });
       print('Failed to get platform version.');
     }
   }
 
-  shareApp(BuildContext context, String text, String share) {
-    final RenderBox box = context.findRenderObject();
+  shareApp(BuildContext context, String? text, String share) {
+    final RenderBox box = context.findRenderObject() as RenderBox;
     Share.share(share,
         subject: text,
         sharePositionOrigin: box.localToGlobal(Offset.zero) & box.size);
@@ -1324,15 +1342,15 @@ class _HomeScreen extends State<HomeScreen>
 }
 
 class WebViewScreen extends StatefulWidget {
-  final GlobalKey webKey;
-  final String path;
-  final int pos;
-  final Settings settings;
-  final InAppWebViewController webViewController;
-  final Stream<int> stream;
+  final GlobalKey? webKey;
+  final String? path;
+  final int? pos;
+  final Settings? settings;
+  final InAppWebViewController? webViewController;
+  final Stream<int>? stream;
 
   WebViewScreen({
-    Key key,
+    Key? key,
     this.path,
     this.webKey,
     this.pos,
@@ -1352,13 +1370,13 @@ class _WebViewScreen extends State<WebViewScreen>
   @override
   bool get wantKeepAlive => true;
 
-  InAppWebViewController _webViewController;
+  InAppWebViewController? _webViewController;
   String url = "";
-  PullToRefreshController pullToRefreshController;
+  PullToRefreshController? pullToRefreshController;
   static GlobalKey<ScaffoldState> _scaffoldKey = new GlobalKey<ScaffoldState>();
 
   List<StreamSubscription<Position>> webViewGPSPositionStreams = [];
-  bool isLoading;
+  late bool isLoading;
 
   final Set<Factory<OneSequenceGestureRecognizer>> _gSet = [
     Factory<VerticalDragGestureRecognizer>(
@@ -1369,7 +1387,7 @@ class _WebViewScreen extends State<WebViewScreen>
 
   @override
   void initState() {
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance!.addObserver(this);
     super.initState();
     isLoading = true;
 
@@ -1409,9 +1427,9 @@ class _WebViewScreen extends State<WebViewScreen>
   @override
   void didUpdateWidget(covariant WebViewScreen oldWidget) {
     if (oldWidget.path != widget.path) {
-      _webViewController.loadUrl(
+      _webViewController!.loadUrl(
         urlRequest: URLRequest(
-          url: Uri.parse(widget.path),
+          url: Uri.parse(widget.path!),
         ),
       );
     }
@@ -1437,7 +1455,7 @@ class _WebViewScreen extends State<WebViewScreen>
 
   @override
   void dispose() {
-    WidgetsBinding.instance.removeObserver(this);
+    WidgetsBinding.instance!.removeObserver(this);
     webViewGPSPositionStreams.forEach(
         (StreamSubscription<Position> _flutterGeolocationStream) =>
             _flutterGeolocationStream.cancel());
@@ -1455,7 +1473,7 @@ class _WebViewScreen extends State<WebViewScreen>
           Expanded(
             child: InAppWebView(
               // contextMenu: contextMenu,
-              initialUrlRequest: URLRequest(url: Uri.parse(widget.path)),
+              initialUrlRequest: URLRequest(url: Uri.parse(widget.path!)),
               gestureRecognizers: _gSet,
               initialOptions: InAppWebViewGroupOptions(
                   crossPlatform: InAppWebViewOptions(
@@ -1464,8 +1482,8 @@ class _WebViewScreen extends State<WebViewScreen>
                     useOnDownloadStart: true,
                     mediaPlaybackRequiresUserGesture: false,
                     userAgent: Platform.isAndroid
-                        ? widget.settings.userAgent.valueAndroid
-                        : widget.settings.userAgent.valueIOS,
+                        ? widget.settings!.userAgent!.valueAndroid!
+                        : widget.settings!.userAgent!.valueIOS!,
                   ),
                   android: AndroidInAppWebViewOptions(
                     useHybridComposition: true,
@@ -1473,7 +1491,7 @@ class _WebViewScreen extends State<WebViewScreen>
                   ios: IOSInAppWebViewOptions(
                     allowsInlineMediaPlayback: true,
                   )),
-              pullToRefreshController: widget.settings.pullRefresh == "1"
+              pullToRefreshController: widget.settings!.pullRefresh == "1"
                   ? pullToRefreshController
                   : null,
               onWebViewCreated: (InAppWebViewController controller) {
@@ -1495,7 +1513,8 @@ class _WebViewScreen extends State<WebViewScreen>
                       switch (action) {
                         case "clearWatch":
                           _geolocationClearWatch(parseInt(
-                              geolocationData['flutterGeolocationIndex'] ?? 0));
+                              geolocationData['flutterGeolocationIndex'] ??
+                                  0)!);
                           break;
 
                         case "getCurrentPosition":
@@ -1511,7 +1530,7 @@ class _WebViewScreen extends State<WebViewScreen>
                           _geolocationWatchPosition(
                               parseInt(
                                   geolocationData['flutterGeolocationIndex'] ??
-                                      0),
+                                      0)!,
                               PositionOptions()
                                   .from(geolocationData['option'] ?? null));
                           break;
@@ -1527,7 +1546,7 @@ class _WebViewScreen extends State<WebViewScreen>
                     action: PermissionRequestResponseAction.GRANT);
               },
               shouldOverrideUrlLoading: (controller, navigationAction) async {
-                var uri = navigationAction.request.url;
+                var uri = navigationAction.request.url!;
                 print("uri.scheme");
                 print(uri.scheme);
                 if (Platform.isAndroid && ["intent"].contains(uri.scheme)) {
@@ -1573,7 +1592,7 @@ class _WebViewScreen extends State<WebViewScreen>
                 });
               },
               onLoadStop: (controller, url) async {
-                pullToRefreshController.endRefreshing();
+                pullToRefreshController!.endRefreshing();
                 Future.delayed(const Duration(milliseconds: 500), () {
                   _geolocationAlertFix();
                 });
@@ -1593,11 +1612,11 @@ class _WebViewScreen extends State<WebViewScreen>
                 }
               },
               onLoadError: (controller, url, code, message) {
-                pullToRefreshController.endRefreshing();
+                pullToRefreshController!.endRefreshing();
               },
               onProgressChanged: (controller, progress) {
                 if (progress == 100) {
-                  pullToRefreshController.endRefreshing();
+                  pullToRefreshController!.endRefreshing();
                 }
               },
               onUpdateVisitedHistory: (controller, url, androidIsReload) {
@@ -1611,23 +1630,23 @@ class _WebViewScreen extends State<WebViewScreen>
             ),
           )
         ]),
-        (isLoading && widget.settings.loader != "empty")
+        (isLoading && widget.settings!.loader != "empty")
             ? Positioned(
                 top: 0,
                 bottom: 0,
                 right: 0,
                 left: 0,
                 child: Loader(
-                    type: widget.settings.loader,
-                    color: themeProvider.isLightTheme
-                        ? HexColor(widget.settings.loaderColor)
+                    type: widget.settings!.loader,
+                    color: themeProvider.isLightTheme!
+                        ? HexColor(widget.settings!.loaderColor)
                         : themeProvider.darkTheme.primaryColor))
             : Container()
       ],
     );
   }
 
-  int parseInt(dynamic value) {
+  int? parseInt(dynamic value) {
     if (value == null) return null;
     if (value is int) return value;
 
@@ -1638,25 +1657,22 @@ class _WebViewScreen extends State<WebViewScreen>
       PositionOptions positionOptions) async {
     PositionResponse positionResponse = PositionResponse();
 
-    int timeout = 30000;
-    if (positionOptions.timeout > 0) timeout = positionOptions.timeout;
+    int? timeout = 30000;
+    if (positionOptions.timeout! > 0) timeout = positionOptions.timeout;
 
     try {
-      Geolocator geolocator = Geolocator()..forceAndroidLocationManager = true;
-      GeolocationStatus geolocationStatus =
-          await geolocator.checkGeolocationPermissionStatus();
+      LocationPermission geolocationStatus =
+          await GeolocatorPlatform.instance.requestPermission();
 
-      if (geolocationStatus == GeolocationStatus.denied ||
-          geolocationStatus == GeolocationStatus.granted ||
-          geolocationStatus == GeolocationStatus.restricted ||
-          geolocationStatus == GeolocationStatus.unknown) {
+      if (geolocationStatus == LocationPermission.always ||
+          geolocationStatus == LocationPermission.whileInUse) {
         positionResponse.position = await Future.any([
-          geolocator.getCurrentPosition(
+          Geolocator.getCurrentPosition(
               desiredAccuracy: (positionOptions.enableHighAccuracy
                   ? LocationAccuracy.best
                   : LocationAccuracy.medium)),
-          Future.delayed(Duration(milliseconds: timeout), () {
-            if (positionOptions.timeout > 0) positionResponse.timedOut = true;
+          Future.delayed(Duration(milliseconds: timeout!), () {
+            if (positionOptions.timeout! > 0) positionResponse.timedOut = true;
             return;
           })
         ]);
@@ -1712,9 +1728,9 @@ class _WebViewScreen extends State<WebViewScreen>
       setTimeout(function(){ _flutterGeolocationAlertFix(); }, 100);
     ''';
 
-    _webViewController.evaluateJavascript(source: javascript);
+    _webViewController!.evaluateJavascript(source: javascript);
 
-    _webViewController.evaluateJavascript(source: """
+    _webViewController!.evaluateJavascript(source: """
       function _flutterGeolocationAlertFix() {
         navigator.geolocation = {};
         navigator.geolocation.clearWatch = function(watchId) {
@@ -1751,7 +1767,7 @@ class _WebViewScreen extends State<WebViewScreen>
 
   void _geolocationClearWatch(int flutterGeolocationIndex) {
     // Stop gps position stream
-    webViewGPSPositionStreams[flutterGeolocationIndex]?.cancel();
+    webViewGPSPositionStreams[flutterGeolocationIndex].cancel();
 
     // remove watcher from list
     webViewGPSPositionStreams.remove(flutterGeolocationIndex);
@@ -1770,11 +1786,11 @@ class _WebViewScreen extends State<WebViewScreen>
       _flutterGeolocationResponse();
     ''';
 
-    _webViewController.evaluateJavascript(source: javascript);
+    _webViewController!.evaluateJavascript(source: javascript);
   }
 
   void _geolocationGetCurrentPosition(
-      int flutterGeolocationIndex, PositionOptions positionOptions) async {
+      int? flutterGeolocationIndex, PositionOptions positionOptions) async {
     PositionResponse positionResponse =
         await getCurrentPosition(positionOptions);
 
@@ -1783,7 +1799,7 @@ class _WebViewScreen extends State<WebViewScreen>
   }
 
   void _geolocationResponse(
-      int flutterGeolocationIndex,
+      int? flutterGeolocationIndex,
       PositionOptions positionOptions,
       PositionResponse positionResponse,
       bool watcher) {
@@ -1795,25 +1811,25 @@ class _WebViewScreen extends State<WebViewScreen>
           ''']({
             coords: {
               accuracy: ''' +
-          positionResponse.position.accuracy.toString() +
+          positionResponse.position!.accuracy.toString() +
           ''',
               altitude: ''' +
-          positionResponse.position.altitude.toString() +
+          positionResponse.position!.altitude.toString() +
           ''',
               altitudeAccuracy: null,
               heading: null,
               latitude: ''' +
-          positionResponse.position.latitude.toString() +
+          positionResponse.position!.latitude.toString() +
           ''',
               longitude: ''' +
-          positionResponse.position.longitude.toString() +
+          positionResponse.position!.longitude.toString() +
           ''',
               speed: ''' +
-          positionResponse.position.speed.toString() +
+          positionResponse.position!.speed.toString() +
           '''
             },
             timestamp: ''' +
-          positionResponse.position.timestamp.millisecondsSinceEpoch
+          positionResponse.position!.timestamp!.millisecondsSinceEpoch
               .toString() +
           '''
           });''' +
@@ -1833,7 +1849,7 @@ class _WebViewScreen extends State<WebViewScreen>
         _flutterGeolocationResponse();
       ''';
 
-      _webViewController.evaluateJavascript(source: javascript);
+      _webViewController!.evaluateJavascript(source: javascript);
     } else {
       // TODO: Return correct error code
       String javascript = '''
@@ -1865,23 +1881,24 @@ class _WebViewScreen extends State<WebViewScreen>
         _flutterGeolocationResponse();
       ''';
 
-      _webViewController.evaluateJavascript(source: javascript);
+      _webViewController!.evaluateJavascript(source: javascript);
     }
   }
 
   void _geolocationWatchPosition(
-      int flutterGeolocationIndex, PositionOptions positionOptions) {
+    int flutterGeolocationIndex,
+    PositionOptions positionOptions,
+  ) {
     // init new strem
-    var geolocator = Geolocator();
-    var locationOptions = LocationOptions(
+    var locationOptions = LocationSettings(
         accuracy: (positionOptions.enableHighAccuracy
             ? LocationAccuracy.best
             : LocationAccuracy.medium),
         distanceFilter: 10);
 
-    webViewGPSPositionStreams[flutterGeolocationIndex] = geolocator
-        .getPositionStream(locationOptions)
-        .listen((Position position) {
+    webViewGPSPositionStreams[flutterGeolocationIndex] =
+        Geolocator.getPositionStream(locationSettings: locationOptions)
+            .listen((Position position) {
       // Send data to each warcher
       PositionResponse positionResponse = PositionResponse()
         ..position = position;
@@ -1892,8 +1909,8 @@ class _WebViewScreen extends State<WebViewScreen>
 
   Future<bool> goBack() async {
     if (_webViewController != null) {
-      if (await _webViewController.canGoBack()) {
-        _webViewController.goBack();
+      if (await _webViewController!.canGoBack()) {
+        _webViewController!.goBack();
         return false;
       } else {
         return showDialog(
@@ -1914,7 +1931,7 @@ class _WebViewScreen extends State<WebViewScreen>
                   ),
                 ],
               ),
-            ) ??
+            ) as FutureOr<bool>? ??
             false;
       }
     }
