@@ -1,6 +1,6 @@
 import 'package:geolocator/geolocator.dart';
 
 class PositionResponse {
-  Position position;
+  Position? position;
   bool timedOut = false;
 }
