@@ -5,7 +5,7 @@ import 'package:http/http.dart' as http;
 class MosqueManager extends ChangeNotifier {
   final sharedPref = SharedPref();
 
-  String mosqueId;
+  String? mosqueId;
 
   Future<void> init() async {
     mosqueId = await sharedPref.read("mosqueId").catchError((e) => null);
@@ -14,7 +14,7 @@ class MosqueManager extends ChangeNotifier {
   }
 
   /// update mosque id in the app and shared preference
-  Future<void> setMosqueId(String id) async {
+  Future<String> setMosqueId(String id) async {
     var url = 'https://mawaqit.net/en/id/$id?view=desktop';
 
     var value = await http.get(Uri.parse(url));
@@ -25,7 +25,7 @@ class MosqueManager extends ChangeNotifier {
       mosqueId = id;
       sharedPref.save('mosqueId', mosqueId);
       notifyListeners();
-      return mosqueId;
+      return mosqueId!;
     }
   }
 }
