@@ -54,7 +54,7 @@ class _QrScanScreen extends State<QrScanScreen> {
     // TODO: implement build
 
     return Scaffold(
-      appBar: _renderAppBar(context, settings),
+      appBar: _renderAppBar(context, settings) as PreferredSizeWidget?,
       body: SafeArea(
         child: Column(
           children: <Widget>[
@@ -82,7 +82,7 @@ class _QrScanScreen extends State<QrScanScreen> {
                 child: Padding(
                     padding: EdgeInsets.all(13),
                     child: Image.network(
-                      settings.logoHeaderUrl,
+                      settings.logoHeaderUrl!,
                     )),
               ),
             ),
@@ -90,7 +90,7 @@ class _QrScanScreen extends State<QrScanScreen> {
               padding: EdgeInsets.only(left: 0, top: 10, right: 0, bottom: 10),
               child: Center(
                   child: Text(
-                settings.title,
+                settings.title!,
                 textAlign: TextAlign.center,
                 style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
               )),
@@ -105,7 +105,7 @@ class _QrScanScreen extends State<QrScanScreen> {
                   EdgeInsets.only(left: 20, top: 40, right: 20, bottom: 10),
               child: Center(
                   child: Text(
-                settings.aboutUs,
+                settings.aboutUs!,
                 textAlign: TextAlign.center,
                 style: TextStyle(fontSize: 18, fontWeight: FontWeight.w300),
               )),
@@ -130,7 +130,7 @@ Widget _renderAppBar(context, Settings settings) {
   var themeProvider = Provider.of<ThemeNotifier>(context);
   return AppBar(
       title: Text(
-        I18n.current.about,
+        I18n.current!.about,
         style: TextStyle(
             color: Colors.white, fontSize: 22.0, fontWeight: FontWeight.bold),
       ),
@@ -140,10 +140,10 @@ Widget _renderAppBar(context, Settings settings) {
             begin: Alignment.centerLeft,
             end: Alignment.centerRight,
             colors: <Color>[
-              themeProvider.isLightTheme
+              themeProvider.isLightTheme!
                   ? HexColor(settings.firstColor)
                   : themeProvider.darkTheme.primaryColor,
-              themeProvider.isLightTheme
+              themeProvider.isLightTheme!
                   ? HexColor(settings.secondColor)
                   : themeProvider.darkTheme.primaryColor,
             ],
